Make test runner fault tolerant
Make the test runner more robust by adding some error handling to the
process of parsing and compiling the event streams.

We should now be able to handle both parse errors and suite crashes in a
gracious manner. Both of these errors produce a special type of event:
`testcase_error`.