Print output when test cases crash
We only printed output when we had a test case that ran all the way
through and produced a `testcase_end` event, but which had the `success`
attribute set to `False`. This means that we do not print anything if
the test case crashes and hence does not produce a `testcase_end` event.

We should now print the output for both cases.