Test module arguments
Instead of just running every test in a module, allow modules to parse
command line arguments and do different actions based on it.

This implements two options for test modules; `list` and `run`.

The `list` option simply lists all the test cases in the module.

The `run` option runs the test cases that is supplied in the following
arguments by index.

These options are useful together to allow the test runner to select
which tests should be run in a module, as opposed to always running all
the available tests.