Improve handling of crash at testcase shutdown
Before this commit it was the testmodules' job to emit module_start and
module_end events. This meant that if testmodule crashed there would be
no module_end event, so the testrunner would fake one to keep it
consistent. However if the testmodule crashed on shutdown - after it had
sent the module_end - there would be two module_end events for the same
module. A similar problem could occur if the testmodule crashed before
module_start had been emitted.

This commit makes the testrunner responsible for always emitting
module_start and module_end events, this way any problem with crashing
testmodule is avoided. The syntesized crash event is also changed to
'module_crash' instead of 'testcase_crash' to make it clearer that
crash may happen outside of a running testcase.