Add --show-output option
As can be guessed from name the new option makes the testrunner show
output also for non-failing testcases.

The output is printed as it appears. To make it easier to understand
where output comes from an extra line with test name is printed
before first output for tests that has output.