testrunner: Add possibility to add extra valgrind options
Extra options to be passed to valgrind can be specified using
`--valgrind-opt` command line argument. It can be specified
multiple times to add multiple options. E.g:

 $ testrunner --valgrind --valgrind-opt "--leak-check=full" --valgrind-opt "--show-leak-kinds=all" mytestmodule

When there is an intermediate buildsystem that makes it hard to pass
options directly to testrunner there is an environment variable that
can be used: `SCU_VALGRIND_OPTIONS`. E.g:

 $ SCU_VALGRIND_OPTIONS="--leak-check=full --show-leak-kinds=all" make tests-valgrind