Count per-test valgrind leaks and show in report
When libscu-c is compiled with valgrind support, add ability to count
leaked memory based on SCU_VALGRIND_LEAK_LEVEL set using memcheck's
own definitions:
    [0 (or not defined) - none, 1 - leaked, 2 - dubious]

VALGRIND_DO_QUICK_LEAK_CHECK is used as it corresponds to
'--leak-check=summary'. Since we are not interested in the specifics
of where the leaks are this should suffice.

Fail test cases where leaks of the set level are present.