Add `--timestampfile` option to allow only running changed testmodules
By specifying a file in `--timestampfile` option SCU will only run the
testcase modules that have an mtime that is newer than mtime of the
specified file. If file is missing all modules are run. After SCU has
finished running all modules it updates the mtime of the timestamp file.