Simplify SCU_TEST/SCU_TAGS macro interaction
The only thing that can be specified in the varargs section of
SCU_TEST is SCU_TAGS (which doesn't really need to be a macro, but for
readability it is nice).

By letting the `{}` be emitted directly in SCU_TEST outside of
SCU_TAGS there are two benefits; there will be no missing field
initializers, so there is no need to ignore that warning, and the
designated inintializer can be removed, making it work in C++ too.