Ensure failures to initiate testmodule is detected
If the testmodule crashed (or didn't start at all) when listing which
tests it contains the module was just skipped as if all tests in it had
been excluded.

Normally libscu-c has the main() function and doesn't run any user code
when listing tests. But a common scenario where this could happen is if
the testcase links with a shared library, and that isn't available in
library path when running the test.

This adds handling so that failed modules are detected and error is
propagated accordingly.